feat: add os check

+++ install.sh
@@ -1,27 +1,38 @@
-snap install ghostty --classic
+#!/bin/bash
 
-# add fish to valid login shells
-command -v fish | sudo tee -a /etc/shells
+# Check if the OS is Ubuntu
+if [ -f /etc/os-release ] && grep -iq "ubuntu" /etc/os-release; then
+  # Run commands for Ubuntu OS
 
-# add fish as default shell
-sudo chsh -s $(which fish) $USER
+  # Install ghostty via snap
+  snap install ghostty --classic
 
-# setup rust
-echo "Installing rust..."
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+  # Add fish to valid login shells
+  command -v fish | sudo tee -a /etc/shells
 
-# Cargo packages
-cargo install cargo-binstall
-cargo binstall eza zellij bat bob-nvim kanata just starship zoxide mise
+  # Set fish as default shell
+  sudo chsh -s $(which fish) $USER
 
-# Mise packages
-mise use go node python@3.12 -g
+  # Setup rust
+  echo "Installing rust..."
+  curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 
-# Go packages
-go install github.com/jesseduffield/lazygit@latest
+  # Install Cargo packages
+  cargo install cargo-binstall
+  cargo binstall eza zellij bat bob-nvim kanata just starship zoxide mise
 
-# stow dotfiles
-stow fish
-stow kanata
-stow nvim
-stow zellij
+  # Install Mise packages
+  mise use go node python@3.12 -g
+
+  # Install Go packages
+  go install github.com/jesseduffield/lazygit@latest
+
+  # Stow dotfiles
+  stow fish
+  stow kanata
+  stow nvim
+  stow zellij
+
+else
+  echo "This script is only compatible with Ubuntu."
+fi
